Work on Add/Modify/Delete Event functionality.
* Create event manager window in MainWindow thread and handle event window
quit message there.
* Disable currently not in use controls in Event Window. Move the logic
into a separate disable control function.
* Replace separate start and end pop-up calendar message with a single
pop-up calendar message and add a bool in order to depict the message
source in Event Window.
* Disable start and end time text controls on AllDay checkbox toggle in
EventWindow. Remove the time text for now until a time control is
implemented.
* Implement EventWindow::SetEvent() functionality which takes in an object
of type Event, a BList containing all events and the index of the event
in the list. It populates all controls based on the event data passed.
* Implement EventWindow::SetEventDate() to set start and end date for
new events. The currently selected date in calendar view is passed to
it. A dummy time is used for now until a time control is implemented.
* Display start/end date and time as per user locale preferences in
EventWindow.
* Implement functionality to save and delete event in EventWindow.
* Implement SidePanelView::GetSelectedDate() to get the currently selected
date in Calendar View.