Differentiate between "canceling" and "deleting"
Now two menu options are shown― one for canceling an event, which will
mark it as "cancelled" (changing the list item's font-face to strikeout
and disabling notification), and another for deletion, which will send
it to the system Trash folder.

The structure of events' status has changed to be a little more
flexible, and the way the file attribute Event:Status is parsed has been
changed (to be a list, e.g., "Cancelled, Notified"). And thanks
to the new dichotomy between "Cancelled" and "Deleted", all previously
deleted events have to be migrated over, which is done on the first
start-up after this update.