Support for "Hiding" events
Events can now be "hidden": undisplayed in Calendar without actually
deleting the file.

The way menu items of and the way hidden/deleted events are filtered out
of Calendar have been tweaked to lay the groundwork for a new DayView
mode or setting that allows management of deleted/hidden events (with the
ignoreHidden parameter of GetEventsOf*() methods).