Event Window tweaks, better opening by file
Adds to the Event Manager window a "Status" box for toggling
whether or not an event is cancelled/hidden.

Also better handles editing events opened directly by their file― all
edits remain with the selected file, even if they otherwise wouldn't
have (e.g., one copy of an event in trash and one not).

Events can also be "Undeleted" from the Event Manager window, which
restores them to their deleted path or to the events directory.