Redesign main window around tabbed events list
Replaces the toolbar with a tabbed events list, "EventTabView,"
which is a renamed and modified DayView.

A "mode" and "view" distinction is made― instead of Day, Week,
and Agenda all being views, the latter is considered a mode,
which can be toggled regardless of the currently selected view.
Views now include a "Month" tab.